replace DataTable with ListView due to performance
The DataTable widget is not made to display more than ~100 rows.
I tested it with 1000 rows which made the app freeze for 10 seconds
before actually displaying the table. So, I replaced the DataTable by
a ListView using the builder method which automatically renders only
those elements which are currently visible.