add UploadProfile as user-defined object metadata
During an upload, the currently enabled UploadProfile is serialized as
JSON and stored in a separate file '_metadata.json'. This file is
uploaded in addition to the actual log files. With this commit, the
properties of the UploadProfile are also assigned as user-defined
object metadata (=S3 feature) to each object / file during upload.