don't show size in bytes of directories
Don't show the size of directories because the value might be mistaken.
The size of a directory itself on Linux is ~4KB and ~0B on Windows. A
user might mistake this size as the sum of all  file sizes which are
contained in the given directory - which it is not.