replace file_picker by native Dart implementation
Replace the usage of the package "file_picker" by our own native
implementation in Dart. The problem with "file_picker" is
that it relies on Go-Flutter and uses the Go package "dlgs"
to show file selection dialogs. However, the builds with
go-hover tend to be buggy and crash very often. The goal is
to remove Go-Flutter / hover completely from this project
to make the resulting build more stable.